src/requirements.txt
google-cloud-pubsub
functions-framework